Deétection si usrager est log in dans les controler cellier et bouteidlle...sinon retourne a la vue login

+++ app/Http/Controllers/BouteilleController.php
@@ -22,26 +22,19 @@ class BouteilleController extends Controller
     public function index(Request $request, $id)
     {
 
-        Auth::check();
-        $id_usager = Auth::id();
+        if(Auth::check()){
+            $id_usager = Auth::id();
 
-        $bouteilles = DB::table('vino__cellier_has_vino__bouteille')
-            ->select('*')
-            ->join('vino__bouteille_personalize', 'vino__bouteille_id', '=', 'vino__bouteille_personalize.id')
-            ->join('vino__cellier', 'vino__cellier_id', '=', 'vino__cellier.id')
-            ->where('vino__cellier_id', $id)
-            ->get();
+            $bouteilles = DB::table('vino__cellier_has_vino__bouteille')
+                ->select('*')
+                ->join('vino__bouteille_personalize', 'vino__bouteille_id', '=', 'vino__bouteille_personalize.id')
+                ->join('vino__cellier', 'vino__cellier_id', '=', 'vino__cellier.id')
+                ->where('vino__cellier_id', $id)
+                ->get();
 
-        //dd($bouteilles);
+            //dd($bouteilles);
 
-       $cellier = Cellier::find($id);
-
-       
-        
-
-
-        //dd($bouteilles);
-
+        $cellier = Cellier::find($id);
 
         return view('bouteille.liste', [
             'bouteilles' => $bouteilles,
@@ -50,6 +43,11 @@ class BouteilleController extends Controller
             'cellier' => $cellier,
             'msg'=> NULL
         ]);
+
+        }else{
+
+            return redirect('/login');
+        }
     }
 
     /*
@@ -57,21 +55,25 @@ class BouteilleController extends Controller
     */
     public function nouveau(Request $request, $id)
     {
-        Auth::check();
-        $id_usager = Auth::id();
-       
-        //Liste des bouteille  
-         $bouteillesSAQ = DB::table('vino__bouteille')
-         ->get();
-
-        //cellier impliquer
-        $cellier = Cellier::find($id);
+        if(Auth::check()){
+            $id_usager = Auth::id();
         
-        //vue des bouteille du catalogue
-        return view('bouteille.nouveau', [
-            'bouteillesSAQ' => $bouteillesSAQ, //pour la rechercher
-            'cellier' => $cellier
-        ]);
+            //Liste des bouteille  
+            $bouteillesSAQ = DB::table('vino__bouteille')
+            ->get();
+
+            //cellier impliquer
+            $cellier = Cellier::find($id);
+            
+            //vue des bouteille du catalogue
+            return view('bouteille.nouveau', [
+                'bouteillesSAQ' => $bouteillesSAQ, //pour la rechercher
+                'cellier' => $cellier
+            ]);
+        }else{
+
+            return redirect('/login');
+        }
     }
 
     /*
@@ -79,42 +81,45 @@ class BouteilleController extends Controller
     */
     public function creer(Request $request)
     {
-        Auth::check();
-        $id_usager = Auth::id();
+        if(Auth::check()){
+            $id_usager = Auth::id();
+            
+            //TODO validate data
+            //$this->validateBouteille($request);
+
+            $quantite = Request::get('quantite');
+            $id_cellier = Request::get('id_cellier');
+
+            //dd($quantite);
+
+            //Ajout de la bouteille dans vin personalize
+            //TODO check duplication//
+            $bouteille = BouteillePersonalize::create(Request::except(['quantite', 'id_cellier']));
+
+            //Ajout de la bouteille dans le cellier 
+            $idBouteille = $bouteille->id;
+            $request2 = [
+                'vino__cellier_id'   => $id_cellier,
+                'vino__bouteille_id' => $idBouteille,
+                'quantite' => $quantite
+            ];
+
+            CelliersBouteilles::create($request2);
+
+            $bouteilles = DB::table('vino__cellier_has_vino__bouteille')
+                ->join('vino__bouteille_personalize', 'vino__bouteille_id', '=', 'vino__bouteille_personalize.id')
+                ->join('vino__cellier', 'vino__cellier_id', '=', 'vino__cellier.id')
+                ->where('vino__cellier_id', $id_cellier)
+                ->get();
         
-        //TODO validate data
-        //$this->validateBouteille($request);
+            //Redirect vers la liste des bouteille du cellier avec un message de succès
+            return redirect()
+            ->route('bouteille.liste', [ 'id' => $id_cellier] )
+            ->withSuccess('Vous avez ajouter la bouteille '.$bouteille->nom.'!');
+        }else{
 
-        $quantite = Request::get('quantite');
-        $id_cellier = Request::get('id_cellier');
-
-        //dd($quantite);
-
-        //Ajout de la bouteille dans vin personalize
-        //TODO check duplication//
-        $bouteille = BouteillePersonalize::create(Request::except(['quantite', 'id_cellier']));
-
-        //Ajout de la bouteille dans le cellier 
-        $idBouteille = $bouteille->id;
-        $request2 = [
-            'vino__cellier_id'   => $id_cellier,
-            'vino__bouteille_id' => $idBouteille,
-            'quantite' => $quantite
-        ];
-
-        CelliersBouteilles::create($request2);
-
-        $bouteilles = DB::table('vino__cellier_has_vino__bouteille')
-            ->join('vino__bouteille_personalize', 'vino__bouteille_id', '=', 'vino__bouteille_personalize.id')
-            ->join('vino__cellier', 'vino__cellier_id', '=', 'vino__cellier.id')
-            ->where('vino__cellier_id', $id_cellier)
-            ->get();
-    
-        //Redirect vers la liste des bouteille du cellier avec un message de succès
-        return redirect()
-        ->route('bouteille.liste', [ 'id' => $id_cellier] )
-        ->withSuccess('Vous avez ajouter la bouteille '.$bouteille->nom.'!');
-
+            return redirect('/login');
+        }
     }
 
 
@@ -125,24 +130,28 @@ class BouteilleController extends Controller
     {
       
         
-       //dd($id);
+        if(Auth::check()){
 
-        $bouteille = BouteillePersonalize::findOrFail($idVin)
-                                            ->select('*')
-                                            ->join('vino__cellier_has_vino__bouteille', 'vino__bouteille_personalize.id', '=', 'vino__cellier_has_vino__bouteille.vino__bouteille_id')
-                                            ->where('id', $idVin)
-                                            ->first();
- 
+            $bouteille = BouteillePersonalize::findOrFail($idVin)
+                                                ->select('*')
+                                                ->join('vino__cellier_has_vino__bouteille', 'vino__bouteille_personalize.id', '=', 'vino__cellier_has_vino__bouteille.vino__bouteille_id')
+                                                ->where('id', $idVin)
+                                                ->first();
+    
 
-       $cellier = Cellier::find($idCellier);
-       
-        //dd($bouteille);
+            $cellier = Cellier::find($idCellier);
+        
+            //dd($bouteille);
 
             return view('bouteille.edit', [
                 'bouteille' => $bouteille,
                 'cellier' => $cellier
                 
             ]);
+        }else{
+
+            return redirect('/login');
+        }
     }
 
     /**
@@ -150,19 +159,23 @@ class BouteilleController extends Controller
      */
     public function update(Request $request, $idVin, $idCellier)
     {
-        //dd($id);
-        $this->validateBouteille($request);
-        $request = Request::all();
+        if(Auth::check()){
+            $this->validateBouteille($request);
+            $request = Request::all();
 
-        $bouteille = BouteillePersonalize::findOrFail($idVin)->update($request);
-        //dd($bouteille);
+            $bouteille = BouteillePersonalize::findOrFail($idVin)->update($request);
+            //dd($bouteille);
 
-        
+            
 
-        // Retourne au formulaire
-        return redirect()
-            ->route('bouteille.liste', [ 'id' => $idCellier] )
-            ->withSuccess('La modification a réussi!');
+            // Retourne au formulaire
+            return redirect()
+                ->route('bouteille.liste', [ 'id' => $idCellier] )
+                ->withSuccess('La modification a réussi!');
+            }else{
+
+                return redirect('/login');
+            }
     }
 
 
@@ -171,28 +184,25 @@ class BouteilleController extends Controller
      */
     public function supprime(Request $request, $idVin, $idCellier)
     {
-        //dd($idVin);
-        $id =(int)$idVin;
-        // TODO lier usager à ses bouteille...
-        $bouteille = BouteillePersonalize::findOrFail($id);
-        
-        
-        $bouteille->delete();
+        if(Auth::check()){
+            $id =(int)$idVin;
+            
+            $bouteille = BouteillePersonalize::findOrFail($id);
+            
+            
+            $bouteille->delete();
 
-        //--------TODO 
 
-        /*supprimer bouteille/ceillier
-        /* model=CelliersBouteilles 
-
-        --------------*/
-
-        // Retourne au formulaire
-        return redirect()
-            ->route('bouteille.liste', [ 'id' => $idCellier] )
-            ->withSuccess("Vous avez supprimer la bouteille  {$bouteille->nom}  !");
+            // Retourne au formulaire
+            return redirect()
+                ->route('bouteille.liste', [ 'id' => $idCellier] )
+                ->withSuccess("Vous avez supprimer la bouteille  {$bouteille->nom}  !");
 
         
-       
+        }else{
+
+            return redirect('/login');
+        }
 
     }
 
@@ -231,14 +241,14 @@ class BouteilleController extends Controller
        //dd($idVin);
         $idCellier = Request::get('idCellier');
         $quantite = Request::get('quantite');
-    dd($idCellier, $idVin, $quantite);
+        //dd($idCellier, $idVin, $quantite);
 
         $updated = CelliersBouteilles::where('vino__bouteille_id', $idVin)
                                         ->limit(1)
                                         ->update(['quantite' => $quantite]); 
 
         
-        dd($updated);
+        //dd($updated);
 
 
        // return json_encode($quantite);
